requirements.txt
flask
flask_session
flask_mysqldb
boto3
python-dotenv
